Update simplify MojitoCore internals for new module strategy.
Rather than having and explicit class function that instantiate types for the
source the backend will instead instatiate the source objects itself based on
types that it discovers through the module loading code.

When instantiating the objects the MojitoCore object will be passed in at
construct time. The sources can then query this with mojito_source_get_core.