Switch the core to use MojitoServiceProxy
Rather than lazily instantiating the MojitoService objects from the backend
upon their addition into a view we instantiate the proxy object on module
load. We then add the proxy object into the view which behaves just like the
real service except that the real service gets instantiated on when needed.