Make MojitoServiceProxy implement the MojitoServiceIface
This interface is the interface generated by dbus-glib from the MojitoService
interface definition file. We add the interface to the MojitoServiceProxy
object since this is the object that we are going to put onto the bus.

When it has dbus methods called on it possibly instantiates the real instance
and then calls the methods on that.