Clarify semantics of service returning NULL for the set
When a service returns NULL for the set this means that the service has been
unable to find out if it has any content i.e. some error occured when it
tried to find out. If the set is NULL then we do not save this into the
cache. This should prevent transient failures from destroying our cache.