Add "avatar-retrieved" signal to MojitoService
When a service has successfully downloaded an avatar after a request it emits
this signal on MojitoService with the local path of the avatar. If it fails
to download the avatar then it should pass NULL in for the path.

This signal is named "avatar-retrieved-internal" to avoid conflicts with the
automatically added signal.