mojito-item: Add API for fetching a remote image and setting key when done
This add mojito_item_request_image_fetch () API call which will attempt to
download the supplied URL to the thumbnail cache. When the download is
complete the given key on the MojitoItem will be set to the local file path.

Having any incomplete downloads will make the MojitoItem unready. When all
downloads are completed the item is ready.