twitter: Don't unref the call. This triggers a bug in librest/libsoup
The bug appears to be that if you unref the call during the callback the
refcount for the call will hit zero after the call has been fired. This causes
the proxy to be unreffed which causes the soup session to be unreffed. This
fails since this is all still in the callback chain for the request's answer.